refactor(callbacks): extract shared infrastructure to _callbacks_shared.py
Step 1 of issue #87 (decompose 7K-line callbacks.py). This PR pulls
the truly cross-cutting infrastructure — caches, design tokens,
layout helpers, basemap constants — into a sibling
``components/_callbacks_shared.py`` module. The 40+ import sites
across ``app.py`` and ``tests/`` continue to resolve unchanged via
an explicit re-import block at the top of ``callbacks.py``.

This is intentionally scoped to ONE module extraction with explicit
test-suite validation as the safety net. The full decomposition
plan (``docs/internal/CALLBACKS_DECOMPOSITION_PLAN.md``) tracks the
per-tab modules and the ``register_callbacks`` split that come in
follow-up PRs.

## What moved

``_callbacks_shared.py`` (236 lines) owns:

- **Caches**: ``_MODEL_CACHE``, ``_PREDICTION_CACHE``, ``_BACKTEST_CACHE``,
  ``_GENERATION_CACHE``, ``_cache_lock``, ``_CACHE_VERSION``
- **Backtest modes**: ``BACKTEST_EXOG_MODES``, ``DEFAULT_BACKTEST_EXOG_MODE``
- **EIA normalization**: ``_EIA_FUEL_MAP``
- **Plotly tokens**: ``PLOT_TEMPLATE``, ``PLOT_LAYOUT``, ``_layout`` helper
- **Color palette**: ``COLORS``, ``_MODEL_BAND_COLORS``
- **Stress threshold**: ``_STRESS_RELIABLE_CEILING``
- **Map basemap tokens**: ``_MAP_LAND_COLOR``, ``_MAP_COASTLINE_COLOR``,
  ``_MAP_SUBUNIT_COLOR``, ``_MAP_AXIS_FONT_COLOR``, ``_MAP_BORDER_COLOR``,
  ``_MAP_COLORSCALE``

## What stays in callbacks.py

Everything else — per-tab helpers, ``register_callbacks(app)``, the
``@app.callback`` decorators, all data-loading paths. They move in
follow-up PRs per the decomposition plan.

## Public-import surface preservation

Used an explicit re-import block (not ``import *``) so ``ruff`` can
statically verify name resolution everywhere downstream. Mutable
singletons (caches) are passed by reference — mutations from inside
``callbacks.py`` and the per-tab modules see the same dict.

## Verification

- callbacks.py: 7,167 → 7,051 lines (will keep dropping as more tabs
  extract; first step is small by design — proves the pattern works
  cleanly with the test suite as the safety net)
- ``ruff check`` + ``ruff format --check`` clean
- All 1,688 existing tests pass — zero modifications to test files
  or to ``app.py``'s import line
- ``app.py:179`` (``from components.callbacks import _BACKTEST_CACHE,
  _MODEL_CACHE, _PREDICTION_CACHE``) continues to resolve correctly
